mang

+++ 1008.cs
@@ -5,8 +5,8 @@ class Program
     public static void Main()
     {
         string[] inputs = Console.ReadLine().Split(' ');
-        int a = int.Parse(inputs[0]);
-        int b = int.Parse(inputs[1]);
+        decimal a = decimal.Parse(inputs[0]);
+        decimal b = decimal.Parse(inputs[1]);
 
         decimal output = a/b;
         Console.WriteLine(output);
